Move assert_fully_folded() into crate::utils::test.
